Admin panel2

See merge request e-commerce-website6150844/e-commerce-website!5

+++ app/Http/Requests/ProductFormRequest.php
@@ -46,11 +46,11 @@ class ProductFormRequest extends FormRequest
             ],
             'original_price' => [
                 'required',
-                'integer',
+                'numeric',
             ],
             'selling_price' => [
                 'required',
-                'integer',
+                'numeric',
             ],
             'quantity' => [
                 'required',
@@ -77,7 +77,6 @@ class ProductFormRequest extends FormRequest
             ],
             'image' => [
                 'nullable',
-            //    'images|mimes:jpeg,png,jpg'
             ],
         ];
     }
